# dist/swagger.yaml
openapi: 3.0.0
info:
  title: APIs Document TeamDUT Blog 
  license:
    name: Contact HuyHo.Xenia
    url: https://www.facebook.com/holehuy23.it
  version: '1.0'
servers:
- url: https://app-xenia.herokuapp.com
  variables: {}
paths:
  /api/users:
    post:
      tags:
      - User
      summary: Create User
      description:
      operationId: CreateUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: admin1@gmail.com
              name: tk1
              password: '123456789'
              passwordConfirmation: '123456789'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - User
      summary: Get Current User
      operationId: GetCurrentUser
      parameters:
      - name: x-refresh
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "{{refreshToken}}"
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "Bearer {{accessToken}}"
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /api/sessions:
    post:
      tags:
      - User
      summary: Create Session
      description: ''
      operationId: CreateSession
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              email: admin1@gmail.com
              password: '123456789'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    delete:
      tags:
      - User
      summary: Delete Session
      description: ''
      operationId: DeleteSession
      parameters:
      - name: x-refresh
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "{{refreshToken}}"
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "Bearer {{accessToken}}"
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - User
      summary: Get Sessions
      operationId: GetSessions
      parameters:
      - name: x-refresh
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "{{refreshToken}}"
      - name: Authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: "Bearer {{accessToken}}"
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /reset-password:
    put:
      tags:
      - Password
      summary: Reset Password
      operationId: ResetPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordSchema'
            example:
              resetLink: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlIjoidXNlciIsImF2YXRhciI6Imh0dHBzOi8vaS5pbWd1ci5jb20vS0R2M25DVC5wbmciLCJyZXNldExpbmsiOiIiLCJfaWQiOiI2MGJlZDU5Mjg0MDI0MzI4ZjM3ZmIxOWYiLCJlbWFpbCI6Imh1eS5ob0B4ZW5pYS50ZWNoIiwibmFtZSI6Imh1eWhvIiwiY3JlYXRlZEF0IjoiMjAyMS0wNi0wOFQwMjoyNzozMC43MzdaIiwidXBkYXRlZEF0IjoiMjAyMS0wNi0wOFQwMjo1MTowNS45NzRaIiwiX192IjowLCJzZXNzaW9uIjoiNjBiZWRiMzcxNjlmNzIzMDMyYjMzY2U2IiwiaWF0IjoxNjIzMTIwNjk1LCJleHAiOjE2MjMxMjI0OTV9.cQOjjrjzixK7YadMagWh1sseM7hCGFUMMcDJLxbPgbo"
              newPassword: "newpassword"
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
components:
  schemas:
    CreateUserRequest:
      title: CreateUserRequest
      required:
      - email
      - name
      - password
      - passwordConfirmation
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 20
        passwordConfirmation:
          type: string
          required: Password much match
      example:
        email: admin1@gmail.com
        name: tk1
        password: '123456789'
        passwordConfirmation: '123456789'
...
    ResetPasswordSchema: 
      title: ResetPasswordRequest
      required:
      - resetLink
      - newPassword
      type: object
      properties:
        resetLink:
          type: string
        newPassword:
          type: string
      example:
        resetLink: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlIjoidXNlciIsImF2YXRhciI6Imh0dHBzOi8vaS5pbWd1ci5jb20vS0R2M25DVC5wbmciLCJyZXNldExpbmsiOiIiLCJfaWQiOiI2MGJlZDU5Mjg0MDI0MzI4ZjM3ZmIxOWYiLCJlbWFpbCI6Imh1eS5ob0B4ZW5pYS50ZWNoIiwibmFtZSI6Imh1eWhvIiwiY3JlYXRlZEF0IjoiMjAyMS0wNi0wOFQwMjoyNzozMC43MzdaIiwidXBkYXRlZEF0IjoiMjAyMS0wNi0wOFQwMjo1MTowNS45NzRaIiwiX192IjowLCJzZXNzaW9uIjoiNjBiZWRiMzcxNjlmNzIzMDMyYjMzY2U2IiwiaWF0IjoxNjIzMTIwNjk1LCJleHAiOjE2MjMxMjI0OTV9.cQOjjrjzixK7YadMagWh1sseM7hCGFUMMcDJLxbPgbo
        newPassword: newpassword
    CreateSessionRequest:
      title: CreateSessionRequest
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: admin1@gmail.com
        password: '123456789'
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...